requirements.txt
Twisted>=16.6.0
attrs>=16.3.0
PyNaCl>=1.0.1
cbor>=1.0.0
